requirements.txt
antiorm==1.2.1
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
behave==1.2.6
blinker==1.7.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.2.1
cryptography==44.0.0
cssselect==1.2.0
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
db==0.1.1
db-sqlite3==0.0.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.1.3
django-allauth==65.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
execnet==2.1.1
executing==2.0.1
fasteners==0.19
filelock==3.16.1
flasgger==0.9.7.1
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.51.0
greenlet==3.0.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mycdp==1.1.0
mysql-connector-python==9.1.0
mysqlclient==2.2.6
nest-asyncio==1.5.8
numpy==1.26.2
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.1.3
parameterized==0.9.0
parse==1.20.2
parse_type==0.6.4
parso==0.8.3
pdbp==1.6.1
pillow==10.3.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.0
PyMySQL==1.1.1
pyngrok==7.1.6
pynose==1.5.3
pyotp==2.9.0
pyparsing==3.1.2
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.3.4
pytest-html==4.0.2
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==15.0
pytest-xdist==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.9.4
rpds-py==0.22.3
sbvirtualdisplay==1.3.1
selenium==4.27.1
seleniumbase==4.33.10
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.23
sqlparse==0.5.2
stack-data==0.6.3
tabcompleter==1.4.0
tenacity==9.0.0
tornado==6.4
traitlets==5.14.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.26.6
wcwidth==0.2.12
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.0.1
wheel==0.45.1
wsproto==1.2.0
zipp==3.20.2


steps:
- name: Install dependencies
  run: pip install -r requirements.txt
